# Speedup train process by reading file in one pass and iterating in parallel over lines.