--- a/tokenizers/src/tokenizer/mod.rs
+++ b/tokenizers/src/tokenizer/mod.rs
@@ -17,7 +17,6 @@ use rayon::prelude::*;
 use std::{
     collections::{HashMap, HashSet},
     fs::File,
-    io::{BufRead, BufReader},
     path::{Path, PathBuf},
 };
 
@@ -559,53 +558,79 @@ impl Tokenizer {
     /// Train a model and replace our current Model, using the given Trainer
     #[allow(clippy::borrowed_box)]
     pub fn train(&mut self, trainer: &Box<dyn Trainer>, files: Vec<String>) -> Result<()> {
-        let progress = ProgressBar::new(100 * files.len() as u64);
-        progress.set_style(
-            ProgressStyle::default_bar()
-                .template("[{elapsed_precise}] {msg:<40!} {wide_bar} {percent:>19!}"),
-        );
-        progress.set_message("Reading files");
+        let len: u64 = files
+            .iter()
+            .map(|filename| File::open(filename).unwrap().metadata().unwrap().len() as u64)
+            .sum();
 
+        let (sender, receiver) = std::sync::mpsc::channel();
+        std::thread::spawn(move || {
+            let progress = ProgressBar::new(len);
+            progress.set_style(
+                ProgressStyle::default_bar()
+                    .template("[{elapsed_precise}] {msg:<40!} {wide_bar} {percent:>19!}"),
+            );
+            progress.set_message(&format!("Reading files ({:.2} Mo)", len / 1_000_000));
+            let mut read = 0;
+            for b in receiver.iter() {
+                read += b as u64;
+                progress.inc(b as u64);
+                if read == len {
+                    break;
+                }
+            }
+            progress.finish();
+        });
         let results = files
-            .into_par_iter()
+            .into_iter()
             .map(|filename| -> Result<HashMap<String, u32>> {
-                let mut words = HashMap::new();
-                let file = File::open(filename)?;
-                let len = file.metadata().map_or(0, |c| c.len());
-                let mut file = BufReader::new(file);
-                let mut prev_prog = 0;
-                let mut read = 0;
-                let mut curr_prog;
+                let buf = std::fs::read_to_string(filename)?;
 
-                let mut buf = String::new();
-                loop {
-                    buf.clear();
-                    // We read new lines using this API instead of the Lines Iterator
-                    // on purpose. We want to keep the `\n` and potential `\r` between each lines
-                    match file.read_line(&mut buf)? {
-                        0 => break,
-                        b => {
-                            let mut normalized = self.do_normalize(NormalizedString::from(&buf))?;
+                let words = buf
+                    .par_split('\n')
+                    .map_with(
+                        sender.clone(),
+                        |sender, line| -> Result<HashMap<String, u32>> {
+                            let mut new_line = line.to_string();
+                            new_line.push_str(&"\n");
+                            let mut words = HashMap::new();
+                            let mut normalized =
+                                self.do_normalize(NormalizedString::from(&new_line))?;
                             let pre_tokenized = self.pre_tokenize(&mut normalized)?;
                             trainer.process_tokens(
                                 &mut words,
                                 pre_tokenized.into_iter().map(|(t, _)| t).collect(),
                             );
 
-                            read += b as u64;
-                            curr_prog = ((read as f64 / len as f64) * 100.0) as u64;
-                            if curr_prog > prev_prog {
-                                progress.inc(curr_prog - prev_prog);
-                                prev_prog = curr_prog;
+                            let b = new_line.len();
+                            sender.send(b)?;
+                            Ok(words)
+                        },
+                    )
+                    .try_fold(
+                        HashMap::new,
+                        |mut acc, ws| -> Result<HashMap<String, u32>> {
+                            {
+                                for (k, v) in ws? {
+                                    *acc.entry(k).or_insert(0) += v;
+                                }
+                            }
+                            Ok(acc)
+                        },
+                    )
+                    .try_reduce(HashMap::new, |mut acc, ws| {
+                        {
+                            for (k, v) in ws {
+                                *acc.entry(k).or_insert(0) += v;
                             }
                         }
-                    }
-                }
+                        Ok(acc)
+                    })?;
 
                 Ok(words)
             })
             .collect::<Vec<_>>();
-        progress.finish();
+        drop(sender);
 
         let mut words = HashMap::new();
         for result in results {