Looking at [YouTokenToMe](https://github.com/VKCOM/YouTokenToMe) it looks like 
tokenizers is still a bit behind in tokenizer training time (BPE) (3-4s vs ~20s on a 
100Mo dataset and it gets worse with size, 10 vs 120s on 500Mo). I'm starting to 
look into it to see what could be done. The first culprit seems to be file loading.
On a 100Mo file, loading everything into memory in one go takes ~30ms on my computer
whereas it takes ~13s to do the "Reading files" step.

There seem to be 3 culprits:
 - First, the read_line method uses a small buffer by
default (8kb) which make tons of syscalls.
 - Second, the parallel iteration occurs
on various files, and not within a single file (most datasets include large files
IMHO).
 - Third, within the line parsing function, lots of string copies are done.

This PR aims at reading the files in one chunk (or at least larger chunks) and 
iterating over lines in parallel instead. The result on my computer on a 100Mo
dataset is going from ~13s to ~7s on the `Reading files` step while maintining 
backward compatibility. Reducing copies is harder but could probably drop the 
time to ~1-2s (~1s is what youtokentome achieves to end up with the full
word count hashmap).

- It will fail on very large files (using read_to_string, *but* it's much faster than previous
iterations. (No bench yet.)
- It should be easy enough to support large files by chunking them at
space boundaries.
- Currently still pretty slow because lots of string copies happen at
train time (Normalizer + PreTokenizer). I also added one to include extra '\n' that
par_lines would remove. No default iterator seems to keep them so we will probably have
to do some custom logic if '\n' need to be kept. It seems Normalizer and PreTokenizer

(Lower bound is indeed 1s : https://gist.github.com/Narsil/833ef43fe4e6169cf54de57733a753b5)