@n1t0 2020-05-22
Hi @Narsil, thank you very much for this. I've been wanting to process lines in parallel for ages but never took the time to do it!

We don't mean to be as fast as YouTokenToMe on the training, and we probably won't ever be. But that is fine! The main reason is that we would like to allow any kind of tokenization, and this kind of flexibility will necessarily mean trading for some speed. YouTokenToMe has been optimized for a specific case and as such, it will obviously perform better on this case.

Now, that being said, I'm all in for everything we can optimize without overcomplicating the API!

I think what takes most of the time is indeed the `Normalizer` and `PreTokenizer` steps. Reading the entire file and counting words is actually really fast (even with `read_line` without parallelization). I'm not sure we can really limit the allocations in the `Normalizer` and `PreTokenizer` since they generally modify the underlying strings. What do you think?

A few things about this PR:
  - I played with it, and it seems to start training around 25% of the `Reading files` steps. I don't know exactly what happens but I keep seeing the progress bar and the progress being made on reading files, while it starts doing everything else at the same time.
 - I'm a bit concerned about reading the files entirely. I feel like it is important to stream them to avoid taking too much memory, especially for users that couldn't use machines with huge amounts of RAM. What do you think?