@Narsil 2020-05-22
> We don't mean to be as fast as YouTokenToMe on the training, and we probably won't ever be. But that is fine! The main reason is that we would like to allow any kind of tokenization, and this kind of flexibility will necessarily mean trading for some speed. YouTokenToMe has been optimized for a specific case and as such, it will obviously perform better on this case.

Yes, that makes sense which is why I use this and not youtokentome, however the sharp difference in speed led me to want to investigate a little. 

> I played with it, and it seems to start training around 25% of the Reading files steps. I don't know exactly what happens but I keep seeing the progress bar and the progress being made on reading files, while it starts doing everything else at the same time.

I use par_lines() and use a channel to send to another thread that keeps updating the progressbar. Maybe the channel buffer gets filled faster than it can be depleted ? So the thread updating that progressbar keeps updating it after it's actually finished ? I used that strategy to avoid sharing the progressbar within par_lines. Oddly enough `progressbar.inc` seems to not mutate itself, so probably better to share directly the progressbar.

> I'm a bit concerned about reading the files entirely. I feel like it is important to stream them to avoid taking too much memory, especially for users that couldn't use machines with huge amounts of RAM. What do you think?

For sure ! I was imagining reading 100Mo or so at a time, only reason it's not done here is because it's actually a bit tricky to get back the next newline character (so we don't cut inside a word). It gets the code a bit more complex.
Other thing I wanted to add was a test for this function, maybe on a smaller data file like `data/big.txt` like the first 100lines or so ? (Maybe adding an opt-out of the progress bar here too).

Another question. Is there any reason why there is no pre_tokenizer by default ? Without it, it changes the speed of this function by quite a bit because the `words` variable basically store the entire dataset in memory instead of doing word counting.