@Narsil 2020-05-23
> For sure ! I was imagining reading 100Mo or so at a time, only reason it's not done here is because it's actually a bit tricky to get back the next newline character (so we don't cut inside a word). It gets the code a bit more complex.

Took me a while to figure out that `par_bridge` was what I was looking for.. Google led me into old rabbit holes :(