@Narsil 2020-05-23
Improved the solution a bit by using a custom Line Iterator that does not remove trailing '\n'  and '\r' ( I copied lines iterator from BufRead trait and removed the pop lines).

I moved the test into benchmark, because it seems Windows does not download the datafiles for tests (it does not read the Makefile). I imagined a unit test would be good, but I admit depending on a file for a unit test is not perfect.

What do you think ? I could remove those benchmarks.