@n1t0 2020-05-23
> Took me a while to figure out that par_bridge was what I was looking for.. Google led me into old rabbit holes :(

Wow! Very good finding, I didn't know about this even though I already looked for something similar multiple times... The resulting implementation is really good!

> I moved the test into benchmark, because it seems Windows does not download the datafiles for tests (it does not read the Makefile). I imagined a unit test would be good, but I admit depending on a file for a unit test is not perfect.

Yes that's perfect, I think it makes more sense as a benchmark. I didn't manage to make it work with windows on the CI, and didn't have a windows handy to debug it at the time. Ideally it should run on windows too indeed.

> I use par_lines() and use a channel to send to another thread that keeps updating the progressbar. Maybe the channel buffer gets filled faster than it can be depleted ?

I found this really surprising and so I ran a few tests. `indicatif` seems to be slowing everything down, and not just a little. See below, running with the last version on this branch:

**100MB while showing progress**
| Train            |   English |   Japanese |   Chinese |   Russian |
|--|--|--|--|--|
| YouTokenToMe     |   3.08772 |    7.71737 |   7.79187 |   2.89366 |
| tokenizers |  16.3958  |   30.0505  |  17.3656  |  17.523   |


**100MB no progress**
| Train            |   English |   Japanese |   Chinese |   Russian |
|--|--|--|--|--|
| YouTokenToMe     |   3.51883 |    8.28085 |   8.13546 |   2.94396 |
| tokenizers |  10.8329  |   15.8105  |   9.88912 |  11.7113  |